Split remote models to a separate file
